chore: modulize authentications in vuex
